Fix split package with org.apache.lucene.index.ShuffleForceMergePolicy
