+++ server/src/main/java/org/elasticsearch/index/engine/ShuffleForcedMergePolicy.java
@@ -6,15 +6,21 @@
  * Side Public License, v 1.
  */
 
-package org.apache.lucene.index;
+package org.elasticsearch.index.engine;
 
+import org.apache.lucene.index.CodecReader;
+import org.apache.lucene.index.FilterMergePolicy;
+import org.apache.lucene.index.LeafReader;
+import org.apache.lucene.index.MergePolicy;
+import org.apache.lucene.index.SegmentCommitInfo;
+import org.apache.lucene.index.SegmentInfos;
+import org.apache.lucene.index.SegmentReader;
 import org.elasticsearch.common.lucene.Lucene;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -67,9 +73,7 @@ public class ShuffleForcedMergePolicy extends FilterMergePolicy {
                 @Override
                 public void setMergeInfo(SegmentCommitInfo info) {
                     // record that this segment was merged with interleaved segments
-                    Map<String, String> copy = new HashMap<>(info.info.getDiagnostics());
-                    copy.put(SHUFFLE_MERGE_KEY, "");
-                    info.info.setDiagnostics(copy);
+                    info.info.addDiagnostics(Map.of(SHUFFLE_MERGE_KEY, ""));
                     super.setMergeInfo(info);
                 }
             });
